Make animations frame rate indepedent aswell as adding a cps counter and a ping/network latency modules. Added whoami mod to run mod list
Animations are based on delta time and hence framerate independent. Cps counter added(works by adding to an arraylist every click(defined by checking when the key is held down whether or not it was held down last frame, hence not the most accurate but works most of the time). Ping is based off the player list(tab list) and hence isn't the most accuriate/up to date. It's gonna be worse than like a serverside /ping but this is the only implementation I could find(thanks kronhud and playerlistping or whatever its called(replaces the connection color icon with the number form of the ping in the tab list)). Add whoami mod because why not